requirements.txt
oauthlib==1.0.3
Pillow==3.1.1
request==0.0.2
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
tweepy==3.5.0
